requirements.txt
crewai>=0.80.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
